Change ProCamCalibration as header only
